det mesta kopplat, förutom font-size för mätvärden...

+++ heimdall_VS/heimdall_VS/Pulse.h
@@ -11,7 +11,8 @@ public:
 	Pulse() = default;
 	~Pulse() = default;
 
-	float calculate();
+	int calculate();
+
 
 private:
 	/*
